feat: apply code review-2
